+++ src/timeline/Timeline.js
@@ -736,10 +736,10 @@ Timeline.prototype.getWindow = function() {
 };
 
 /**
- * Force a repaint of the Timeline. Can be useful to manually repaint when
+ * Force a redraw of the Timeline. Can be useful to manually redraw when
  * option autoResize=false
  */
-Timeline.prototype.repaint = function() {
+Timeline.prototype.redraw = function() {
   var resized = false,
       options = this.options,
       props = this.props,
@@ -845,6 +845,10 @@ Timeline.prototype.repaint = function() {
     // keep repainting until all sizes are settled
     this.repaint();
   }
+
+// TODO: deprecated since version 1.1.0, remove some day
+Timeline.prototype.repaint = function repaint() {
+    throw new Error('Function repaint is deprecated. Use redraw instead.');
 };
 
 /**
@@ -871,8 +875,9 @@ Timeline.prototype._onSelectItem = function (event) {
 
   var newSelection = this.getSelection();
 
-  // if selection is changed, emit a select event
-  if (!util.equalArray(oldSelection, newSelection)) {
+  // emit a select event,
+  // except when old selection is empty and new selection is still empty
+  if (newSelection.length > 0 || oldSelection.length > 0) {
     this.emit('select', {
       items: this.getSelection()
     });
@@ -931,7 +936,7 @@ Timeline.prototype._onAddItem = function (event) {
     this.options.onAdd(newItem, function (item) {
       if (item) {
         me.itemsData.add(newItem);
-        // TODO: need to trigger a repaint?
+        // TODO: need to trigger a redraw?
       }
     });
   }